email-job-agent/requirements.txt
portia-sdk-python[all]
python-dotenv